+++ tests/test_api/Suite/Tools/ContractsManager.py
