+++ tests/test_api/Suite/Tools/Wallet.py
